Alteração do nome do pacote activity para ui e correção de viewHolder
