+++ app/src/main/AndroidManifest.xml
@@ -13,27 +13,26 @@
         android:theme="@style/Theme.MontaigneApp"
         tools:targetApi="31">
         <activity
-            android:name=".SptActivity"
+            android:name=".ui.spt.SptActivity"
             android:exported="false" />
         <activity
-            android:name=".activity.spt.carimboEnsaio.CarimboEnsaiooActivity"
+            android:name=".ui.spt.carimboEnsaio.CarimboEnsaioFragment"
             android:exported="false" />
         <activity
-            android:name=".activity.spt.projeto.ProjetoActivity"
+            android:name=".ui.spt.projeto.ProjetoActivity"
             android:exported="false" />
         <activity
-            android:name=".activity.spt.furo.FuroActivity"
+            android:name=".ui.spt.furo.FuroActivity"
             android:exported="false" />
         <activity
-            android:name=".activity.spt.ensaio.EnsaioActivity"
+            android:name=".ui.spt.ensaio.EnsaioActivity"
             android:exported="false" />
         <activity
-            android:name=".activity.carimboProjeto.CarimboProjetoActivity"
+            android:name=".ui.carimboProjeto.CarimboProjetoFragment"
             android:exported="false" />
         <activity
-            android:name=".activity.home.HomeActivity"
+            android:name=".ui.home.HomeActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.MontaigneApp.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
