+++ app/src/main/java/com/montaigne/montaigneapp/ui/carimboProjeto/CarimboProjetoFragment.java
@@ -1,22 +1,16 @@
-package com.montaigne.montaigneapp.activity.carimboProjeto;
+package com.montaigne.montaigneapp.ui.carimboProjeto;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageButton;
-import android.widget.ImageView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.montaigne.montaigneapp.R;
-import com.montaigne.montaigneapp.activity.AbstractActivity;
-import com.montaigne.montaigneapp.databinding.FragmentCarimboEnsaioBinding;
 import com.montaigne.montaigneapp.databinding.FragmentCarimboProjetoBinding;
 
 import java.util.HashMap;
@@ -34,22 +28,21 @@ public class CarimboProjetoFragment extends Fragment {
         super.onCreate(savedInstanceState);
 
         CarimboProjetoVM viewModel = new ViewModelProvider(this).get(CarimboProjetoVM.class);
-
     }
+
     @Nullable
     @Override
     public View onCreateView (@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState){
         super.onCreateView(inflater, container, savedInstanceState);
-
         binding = FragmentCarimboProjetoBinding.inflate(inflater, container, false);
         fields.put ("Tecnico",  binding.editTextTecnico);
-       fields.put ("Empresa",  binding.editTextEmpresa);
-       fields.put ("Contato",  binding.editTextContato);
-       fields.put ("Cliente",  binding.editTextCliente);
-       fields.put ("LocalObra",  binding.editTextLocalObra);
-       fields.put ("QuantidadeFuros",  binding.editTextQuantidadeFuros);
-       binding.buttonContinueCarimbo.setOnClickListener(v -> viewModel.continuarCarimboButtonListener (v, fields));
-       binding.imageButtonHome.setOnClickListener(viewModel::homeButtonListener);
+        fields.put ("Empresa",  binding.editTextEmpresa);
+        fields.put ("Contato",  binding.editTextContato);
+        fields.put ("Cliente",  binding.editTextCliente);
+        fields.put ("LocalObra",  binding.editTextLocalObra);
+        fields.put ("QuantidadeFuros",  binding.editTextQuantidadeFuros);
+        binding.buttonContinueCarimbo.setOnClickListener(v -> viewModel.continuarCarimboButtonListener (v, fields));
+        binding.imageButtonHome.setOnClickListener(viewModel::homeButtonListener);
 
         return binding.getRoot();
     }
