+++ app/src/main/java/com/montaigne/montaigneapp/ui/home/HomeActivity.java
@@ -1,4 +1,4 @@
-package com.montaigne.montaigneapp.activity.home;
+package com.montaigne.montaigneapp.ui.home;
 
 import android.os.Bundle;
