+++ app/src/main/java/com/montaigne/montaigneapp/ui/spt/furo/FuroActivity.java
@@ -1,4 +1,4 @@
-package com.montaigne.montaigneapp.activity.spt.furo;
+package com.montaigne.montaigneapp.ui.spt.furo;
 
 import android.os.Bundle;
 import android.widget.Button;
@@ -9,7 +9,7 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.montaigne.montaigneapp.R;
-import com.montaigne.montaigneapp.activity.AbstractActivity;
+import com.montaigne.montaigneapp.ui.AbstractActivity;
 import com.montaigne.montaigneapp.model.spt.FuroSpt;
 import com.montaigne.montaigneapp.model.spt.ProjetoSpt;
