+++ app/src/main/java/com/montaigne/montaigneapp/ui/spt/projeto/FurosAdapter.java
@@ -1,4 +1,4 @@
-package com.montaigne.montaigneapp.activity.spt.furo;
+package com.montaigne.montaigneapp.ui.spt.projeto;
 
 import android.content.Intent;
 import android.view.LayoutInflater;
@@ -12,45 +12,33 @@ import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.montaigne.montaigneapp.R;
-import com.montaigne.montaigneapp.activity.spt.ensaio.EnsaioActivity;
-import com.montaigne.montaigneapp.model.spt.AmostraSpt;
+import com.montaigne.montaigneapp.ui.spt.furo.FuroActivity;
+import com.montaigne.montaigneapp.model.spt.FuroSpt;
 import com.montaigne.montaigneapp.model.spt.ProjetoSpt;
 
 import java.util.List;
 
-public class AmostraAdapter extends RecyclerView.Adapter<AmostraAdapter.ViewHolder>{
-    private int idFuro;
-    private ProjetoSpt projetoSpt;
-    private List<AmostraSpt> amostras;
-
-    public void setAmostras(List<AmostraSpt> amostras) {
-        this.amostras = amostras;
-    }
-
-    public void setIdFuro(int id) {
-        this.idFuro = id;
-    }
-
-    public void setProjetoSpt(ProjetoSpt projetoSpt) {
-        this.projetoSpt = projetoSpt;
-    }
+public class FurosAdapter extends RecyclerView.Adapter<FurosAdapter.ViewHolder> {
+    protected ProjetoSpt projetoSpt;
+    protected List<FuroSpt> furos;
 
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.adapter_projeto, parent, false);  // ver se o layout deve ser alterado
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.adapter_projeto, parent, false);
         return new ViewHolder(view);
     }
 
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        holder.textView.setText("Amostra " + (position + 1));
-        int idAmostra = position;
+        // TODO: verificar se vai ser realmente o código que será exibido e como gerá-lo
+        holder.textView.setText(furos.get(position).getCodigo());
+
+        int id = position;
 
         holder.cardView.setOnClickListener(v -> {
-            Intent intent = new Intent(v.getContext(), EnsaioActivity.class);
-            intent.putExtra("idAmostra", idAmostra);
-            intent.putExtra("idFuro", idFuro);
+            Intent intent = new Intent(v.getContext(), FuroActivity.class);
+            intent.putExtra("idFuro", id);
             intent.putExtra("projetoSpt", projetoSpt);
 
             v.getContext().startActivity(intent);
@@ -59,7 +47,19 @@ public class AmostraAdapter extends RecyclerView.Adapter<AmostraAdapter.ViewHold
 
     @Override
     public int getItemCount() {
-        return amostras.size();
+        if (furos == null) {
+            return 0;
+        }
+        return furos.size();
+    }
+
+    // TODO: seria bom pensarmos no caso de utilizar um construtor para recebimento desses dados
+    public void setFuros(List<FuroSpt> furos) {
+        this.furos = furos;
+    }
+
+    public void setProjetoSpt(ProjetoSpt projetoSpt) {
+        this.projetoSpt = projetoSpt;
     }
 
     protected class ViewHolder extends RecyclerView.ViewHolder{
@@ -71,7 +71,6 @@ public class AmostraAdapter extends RecyclerView.Adapter<AmostraAdapter.ViewHold
             super(itemView);
             cardView = itemView.findViewById(R.id.cardView);
             textView = itemView.findViewById(R.id.textFuroName);
-            // todo: garantir que este nome sempre seja um int
             checkBox = itemView.findViewById(R.id.checkBox);
         }
     }
