+++ app/src/main/res/layout/activity_spt.xml
@@ -5,7 +5,7 @@
     android:id="@+id/containerSpt"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.spt.SptActivity" >
+    tools:context=".ui.spt.SptActivity" >
